6.22
UI stuff
 - buttons
 - text & text input
 - ui handler
 - ui manager

+++ 5.24/Camera.cpp
@@ -79,7 +79,7 @@ void Camera::update() {
 	}
 }
 
-void Camera::update(int x, int y) {
+void Camera::update(double x, double y) {
 	if (_is_locked) {
 		return;
 	}
